Integrate ASP.NET Identity
Change DbContext base class to IdentityDbContext (generic class). IdentityUser class defines the base properties that I want to capture from the user.
To use Identity, I needed to configure it in the Startup.cs ConfigureServices. It will use the AppDbContext to store the information. I do not need to write queries. That is already taken care of in an API.

In the configure method, I configure Authentication. It is important that it is placed before the MVC. Components are placed in a sequential order.

I updated my database to use Authentication through Package Manager Console